Move DIL tests for shared & unique pointers into existing dataformatter tests.
