--- lldb/test/API/functionalities/data-formatter/data-formatter-stl/libcxx/dil_shared_ptr/Makefile
@@ -1,6 +1,0 @@
-CXX_SOURCES := main.cpp
-CXXFLAGS_EXTRAS := -std=c++14
-
-USE_LIBCPP := 1
-
-include Makefile.rules
